+++ tests/testthat/test-measures.R
@@ -41,7 +41,7 @@ testthat::test_that("Calculating Summaries Works", {
         MIMS_UNIT = 5.32015536241241)
       , tolerance = 1e-5
     )
-    n_idle = calculate_n_idle(res, epoch = "5 min")
+    n_idle = calculate_n_idle(res, unit = "5 min")
     testthat::expect_equal(
       n_idle$n_idle,
       c(4300L, 28900L, 30000L, 28600L, 30000L,
@@ -55,7 +55,7 @@ testthat::test_that("Calculating Summaries Works", {
       c(AI = 1.93017183478711, SD = 0.104740319672159, MAD = 0.0606835327590105)
       , tolerance = 1e-5
     )
-    n_idle = calculate_n_idle(res, epoch = "5 min")
+    n_idle = calculate_n_idle(res, unit = "5 min")
     testthat::expect_equal(
       n_idle$n_idle,
       c(4300L, 28900L, 30000L, 28600L, 30000L,