+++ NEWS.md
@@ -1,7 +1,7 @@
 # SummarizedActigraphy 0.3.0
 
-- Slot is now `data`  not `data.out` in `read_actigraphy`.
-
+- Slot is now `data` not `data.out` in `read_actigraphy`.
+- All things `epoch` changed to `unit`. 
 
 # SummarizedActigraphy 0.1.0